Add  - Enhance code for support multiple runtime type (ex. ADF , API , CLI etc.)
