+++ src/mod_runtime_api/runtime_api.rs
@@ -6,9 +6,9 @@ use crate::mod_runtime_api::entities::{
     ActivityType, ExFlowRuntimeActivityWebRequest, ExFlowRuntimeWebError, PipelineRunRequest,
     PipelineRunResponse,
 };
-use crate::mod_runtime_cli::adf_runtime::ExFlowRuntimeADFActivityExecutor;
+use crate::mod_runtime_cli::adf_runtime::{ExFlowRuntimeActivityADFParam, ExFlowRuntimeADFActivityExecutor};
 use crate::mod_runtime_cli::interface_runtime::{
-    ExFlowRuntimeActivityADFParam, ExFlowRuntimeActivityExecutor,
+    ExFlowRuntimeActivityExecutor,
 };
 
 type ExFlowWebRuntimeResult<T> = Result<T, ExFlowRuntimeWebError>;
